Fix windows regressions
As I hoped, none of the changes broke windows specifically, beyond just
the untested change to lazily fetched GL functions to avoid having to
require the GL library at runtime which was nasty.

Beyond that, I just needed to support a consistent way for windows to
trigger a "create" event.

Right now both the callback and the window aren't registered in the
eventloop when the WM_CREATE real event is fired, so I couldn't use
that, in the future I'll investigate ways to try to mitigate this but
it's difficult.