Very close
I think the last things that need to be done are uniforms being
refactored into UBOs and then buffers generically. And then something
harder, the texture manager..

This commit actually gets it partially working, it did at least clear
the screen previously. Now the problem is a framebuffer not being
complete, presumably an issue with a bunch of uniform code still
existing