Clear specific depth
This is so incredibly close to working its actually very frustrating!!!
That's the nice way to put it !!!

Anyway the last issue is that "TODO" i made blissfully thinking i might
be able to ignore the explicit depth buffer texture. I believe I need to
do it as the problem is the depth buffer doesn't seem to exist when
using the opengl introspection functions, and of course nothing draws.
At least thats my theory.