Cross platform texture writing
Now queue has a writeTexture method which uses the information from the
texture descriptor so you don't have to provide as much info at runtime.

This behaves like newer rendering APIs, so textures cannot be written to
in other formats for example with automatic conversions.

As a consequence, the texture manager now only accepts rgba8 format
images. But it seems like everything was already using that beyond the
white default texture which was easy to port over.